Move function back into config cjs
